requirements.txt
asgiref==3.8.1
attrs==24.2.0
autoflake==2.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
Django==5.1.1
djangorestframework==3.15.2
flake8==7.1.1
h11==0.14.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
outcome==1.3.0.post0
packaging==24.1
psycopg2==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
